resolved warnings for type-cast

+++ ComGraphics/ComGraphics/Source/SceneLevelOneA.cpp
@@ -353,7 +353,7 @@ void SceneLevelOneA::RenderModelOnScreen(Mesh* mesh, float size, float Rotate, i
     modelStack.LoadIdentity(); //Reset modelStack
     modelStack.Scale(size, size, size);
     modelStack.Translate(x, y, z);
-    modelStack.Rotate(Rotate, rX, rY, rZ);
+	modelStack.Rotate(Rotate, (float)rX, (float)rY, (float)rZ);
 
     RenderMesh(mesh, LightYN);
